backend/ai_service/requirements.txt
opencv-python-headless
numpy
requests
Pillow
